requirements.txt
watchdog==4.0.0
pystray==0.19.5
Pillow==10.2.0
pyinstaller==6.5.0